Improved the performance a bit
Instead of having the code search through every position in the tilemap, it now only checks all tile position within a square of side length 2*bounds. Also, for the future I will save the tilePoslist so that the position of  the tiles does not have to be computed during play time.